[#47359755] Replace lxml with minidom to fix xmlns
The resulting document was missing an `xmlns:vcloud` attribute in the parent
node which was causing vCloud to barf on OVF/OVA imports because it doesn't
understand what `vcloud:` elements are.

lxml doesn't allow you to modify the namespace mapping of a document without
creating a new document and copying all nodes over. Replacing lxml with
minidom, which does permit that, and is in some ways simpler.

As a side-effect the "sub-documents" that we add no longer contain `xmlns`
attributes and `required` appears as `ovf:required` which matches the OVF
that can be subsequently extracted from vCloud's API. It was originally
thought that these would cause problems, but they were red herrings in light
of the issue described above.