Change VMware output from personal to enterprise
This output format is intended for ESX[i] instead of Fusion, which matches
our vCloud destination more closely.

In practice this has very little effect on the resulting OVF. It changes the
default name of the network from `NAT` to `VM Network`. It is meant to allow
customisation of the network name using `hardware.network`, but Boxgrinder's
schema validation doesn't actually allow for this.