[#55141908] Use official packages for VMware Tools
The open source `open-vm-tools` package doesn't allow Guest Customization on
vCloud 5.x - the flag gets disabled when an OVA is uploaded and the checkbox
remains greyed out in the UI, for both the template and any subsequently
provisioned VMs.

This means that newly provisioned VMs (from `vcloud-box-spinner` which uses
instantiate vApp from template) don't set their hostname and static IP
addresses.

The official packages do support customization, to some extent. It doesn't
seem to fix the issue of the option being disabled in the template or newly
provisioned VM. However when a new VM is shutdown, the option can be
changed, and the hostname/network get set correctly on the next boot.

It may be that we need to change `vcloud-box-spinner` to cycle the power on
a new VM before apply any customisations.

http://www.vmware.com/download/packages.html

I wasn't able to get this working on the `vmware` platform-specific
postinstall section. It was bombing out on some libguestfs related errors, I
think -- sorry I've since lost the errors and it takes ages to reproduce. As
a result I've just tacked this onto the base section.